+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- server: Fix page reload with base path
 - webapp: Fix browser based router with base path
 
 ## [1.17.0] - 2024-09-15